fix extensions
